+++ src/main/java/com/colaborativo/modulo_lyd/model/conductor/DatosActualizarConductor.java
@@ -12,7 +12,7 @@ import jakarta.validation.constraints.NotNull;
  */
 public record DatosActualizarConductor(
         @NotNull
-        Integer idConductor,
+        Long idConductor,
         String nombreConductor,
         String licenciaVigente,
         String twicCard,
